+++ modules/ffm-plus/src/main/java/club/doki7/ffm/library/ILibraryLoader.java
@@ -1,17 +1,20 @@
 package club.doki7.ffm.library;
 
+import org.jetbrains.annotations.NotNull;
+
 public sealed interface ILibraryLoader permits
         WindowsLibraryLoader,
-        DefaultLibraryLoader
+        UnixLibraryLoader,
+        JavaSystemLibraryLoader
 {
-    ISharedLibrary loadLibrary(String libName) throws UnsatisfiedLinkError;
+    @NotNull ISharedLibrary loadLibrary(@NotNull String libName) throws UnsatisfiedLinkError;
 
-    default ISharedLibrary loadLibraryFile(String fileName) throws UnsatisfiedLinkError {
-        return loadLibrary(fileName);
-    }
-
-    static ISharedLibrary platformLibraryLoader() {
-        // TODO
-        return null;
+    static ILibraryLoader platformLoader() {
+        String osName = System.getProperty("os.name").toLowerCase();
+        if (osName.contains("windows")) {
+            return WindowsLibraryLoader.INSTANCE;
+        } else {
+            return UnixLibraryLoader.INSTANCE;
+        }
     }
 }