+++ modules/ffm-plus/src/main/java/club/doki7/ffm/library/ISharedLibrary.java
@@ -3,6 +3,7 @@ package club.doki7.ffm.library;
 import club.doki7.ffm.RawFunctionLoader;
 
 public sealed interface ISharedLibrary extends RawFunctionLoader, AutoCloseable permits
-        WindowsLibrary,
-        DefaultLibrary
+        JavaSystemLibrary,
+        UnixLibrary,
+        WindowsLibrary
 {}