+++ modules/ffm-plus/src/main/java/club/doki7/ffm/library/JavaSystemLibraryLoader.java
@@ -1,0 +1,19 @@
+package club.doki7.ffm.library;
+
+import org.jetbrains.annotations.NotNull;
+
+public final class JavaSystemLibraryLoader implements ILibraryLoader {
+    @Override
+    public @NotNull ISharedLibrary loadLibrary(@NotNull String libName) throws UnsatisfiedLinkError {
+        try {
+            System.loadLibrary(libName);
+        } catch (Throwable e) {
+            System.load(libName);
+        }
+        return JavaSystemLibrary.INSTANCE;
+    }
+
+    public static final JavaSystemLibraryLoader INSTANCE = new JavaSystemLibraryLoader();
+
+    private JavaSystemLibraryLoader() {}
+}